Audio is still broken.

+++ VoiceCraft.Server/Systems/EntityEventsSystem.cs
@@ -138,7 +138,7 @@ namespace VoiceCraft.Server.Systems
             {
                 //Only send updates to visible entities.
                 var visibleEntities = entity.VisibleEntities.Where(x => x != entity);
-                var packet = new AudioPacket(entity.Id, timestamp, data.Length, data);
+                var packet = new AudioPacket(entity.Id, timestamp, (ushort)data.Length, data);
                 foreach (var visibleEntity in visibleEntities)
                 {
                     _networkSystem.SendPacket(visibleEntity.NetPeer, packet);
